refactored Upload code into new component
