+++ restapi/stories.go
@@ -78,7 +78,8 @@ func (api api) AddAuthorToStories(stories []models.Story) ([]models.Story, error
 func (api api) ReturnAllStories(w http.ResponseWriter, r *http.Request) render.Renderer {
 	var stories []models.Story
 
-	err := api.database.Where("is_visible = true").Find(&stories).Error
+	err := api.database.Find(&stories).Error
+	// err := api.database.Where("is_visible = true").Find(&stories).Error
 	if err != nil {
 		return rest.ErrInternal(api.logger, err)
 	}
